psycopg mistake update
I changed the psycop on line 34 back to >=2.6.2. I originally changed it, because I was following the suggestions to fixing a problem that was preventing me from deploying the portfolio app successfully to Heroku. It didnt solve the problem so I changed it back to its original state to avoid any more possible errors.